use AC_CHECK_TOOL for libusb-config
AC_PATH_PROG is for build tools while AC_PATH_TOOL is for the host tools.
Since we want a libusb that is for the host machine and not the build
machine, use the TOOL variant.

I don't personally see the point in using the full path either since the
short name will be found in a PATH lookup, so further change the macro
from the PATH variant to the CHECK variant.

Also drop the incorrect third argument of $PATH.  The third argument is
a fall back value, not the path to search.  Ignoring that, $PATH is the
default value anyways, so no need to specify it.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>